[Rename] 파일명 paging 단어 제거
